Refactor member modal types and file names
